+++ moneylink-backend/src/Controllers/HouseholdController/index.js
@@ -3,11 +3,13 @@ const { Household, Bill, User } = require('../../Models');
 const createHousehold = async (req, res) => {
   const { name } = req.body;
 
+  if (!name) return res.status(400).json({ error: 'Missing parameter "name"' });
+
   const household = new Household({
     name: name,
   });
 
-  return await household.save().then((household) => res.status(201).json({ household }));
+  return await household.save().then((household) => res.status(201).json(household));
 };
 
 const getHousehold = async (req, res) => {
@@ -18,9 +20,9 @@ const getHousehold = async (req, res) => {
   return await Household.findById(_id)
     .then((household) => {
       if (!household) return res.status(404).json({ error: 'Household not found' });
-      return res.status(200).json({ household });
+      return res.status(200).json(household);
     })
-    .catch((error) => res.status(400).json({ error }));
+    .catch((error) => res.status(400).json(error));
 };
 
 const updateHousehold = async (req, res) => {
@@ -31,16 +33,18 @@ const updateHousehold = async (req, res) => {
   await Household.findByIdAndUpdate(_id, req.body);
 
   return await Household.findById(_id)
-    .then((household) => res.status(200).json({ household }))
-    .catch((error) => res.status(400).json({ error }));
+    .then((household) => res.status(200).json(household))
+    .catch((error) => res.status(400).json(error));
 };
 
 const deleteHousehold = async (req, res) => {
   const { _id } = req.body;
 
+  if (!_id) return res.status(400).json({ error: 'Missing parameter "id"' });
+
   return await Household.findByIdAndDelete(_id)
-    .then((household) => res.status(200).json({ household }))
-    .catch((error) => res.status(400).json({ error }));
+    .then((household) => res.status(200).json(household))
+    .catch((error) => res.status(400).json(error));
 };
 
 module.exports = {
